CHORE: Refactor roadmap data and remove inline CSS (#175)
* :art: improve structure and refactor

* :art: shift roadamp data to websiteData from page component

* :fire: remove inline css

* update gitignore

* update gitignore

* Rename communitycard.module.css to CommunityCard.module.css

* Rename meetingcard.module.css to MeetingCard.module.css
