+++ .gitignore
@@ -5,7 +5,7 @@ node_modules
 contributors.json
 
 # Production
-/build
+build/
 
 # Generated files
 .docusaurus
