+++ website/src/pages/ProjectCard/card.css
@@ -1,103 +1,101 @@
-
 .card {
-    width: 500px;
+  width: 500px;
+  height: 250px;
+  margin-bottom: 10%;
+  margin-left: 15%;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
+  transition: 0.3s;
+}
+
+.card:hover {
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+}
+.card_container {
+  padding: 10px;
+  min-height: 180px;
+  max-height: 185px;
+  overflow-y: auto;
+}
+.card_container h3 {
+  text-align: center;
+  text-decoration: underline;
+  text-decoration-color: #72e7a3;
+}
+.card_body {
+  width: 100%;
+  height: 100%;
+}
+.card_button {
+  padding: 2%;
+  background-color: white;
+  border: none;
+  font-size: 16px;
+  font-family: Arial, Helvetica, sans-serif;
+}
+.card_button:hover {
+  background-color: #72e7a3;
+  color: white;
+}
+@media screen and (max-width: 966px) {
+  .card {
+    width: 100%;
     height: 250px;
     margin-bottom: 10%;
-    margin-left: 15%;
-    box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2);
     transition: 0.3s;
   }
-  
-.card:hover {
-    box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
 }
-.card_container{
-    padding: 10px;
-    min-height: 180px;
-    max-height: 185px;
-    overflow-y: auto;    
-}
-.card_container h3{
-    text-align: center;
-    text-decoration: underline;
-    text-decoration-color: #72E7A3;
-}
-.card_body{
-    width: 100%;
-    height: 100%;
-}
-.card_button{
-    padding: 2%;
-    background-color: white;
-    border: none;
-    font-size: 16px;
-    font-family:Arial, Helvetica, sans-serif;
-}
-.card_button:hover{
-    background-color: #72E7A3;
-    color: white;
-}
-@media screen and (max-width:966px) {
-    
-    .card{
-        width: 100%;
-        height: 250px;
-        margin-bottom: 10%;
-        box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2);
-        transition: 0.3s;
-    }
-}
-.modal{
-    overflow: auto;
+.modal {
+  overflow: auto;
 }
 .modal-backdrop {
-    position: fixed;
-    height: 100vh;
-    width: 100vw;
-    top: 0;
-    left: 0;
-    background: rgba(0, 0, 0, 0.6);
-  }
-  
+  position: fixed;
+  height: 100vh;
+  width: 100vw;
+  top: 0;
+  left: 0;
+  background: rgba(0, 0, 0, 0.6);
+}
+
 .modal-content-wrapper {
-    display: block;
-    position: fixed;
-    width: 100%;
-    height: 500px;
-    background: #18191A;
-    margin: auto;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    padding: 50px;
-    overflow: auto;
-    color:white;
+  display: block;
+  position: fixed;
+  width: 100%;
+  height: 500px;
+  background: #18191a;
+  margin: auto;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  padding: 50px;
+  overflow: auto;
+  color: white;
 }
-.modal_links{
-    display: flex;
-    align-items: center;
-    margin-left: 45%;
-    font-size: 20px;
-    text-decoration: underline;
+.modal_links {
+  display: flex;
+  align-items: center;
+  margin-left: 45%;
+  font-size: 20px;
+  text-decoration: underline;
 }
-.modal_gh_link{
-    align-items: center;
-    padding: 10px;
+.modal_gh_link {
+  align-items: center;
+  padding: 10px;
 }
-.task_timlines{
-    display: flex;
+.task_timlines {
+  display: flex;
 }
-.status{
-    width: 30%;
-    height: 40%;
-    text-align: justify;
-    text-justify: inter-word;
-    margin-left: 2%;
+.status {
+  width: 30%;
+  height: 40%;
+  text-align: justify;
+  text-justify: inter-word;
+  margin-left: 2%;
 }
-.status h5{
-    text-align: center;
-    text-decoration: underline;
-    text-decoration-color: #72E7A3;
-    text-decoration-thickness: 3px;
+.status h5 {
+  text-align: center;
+  text-decoration: underline;
+  text-decoration-color: #72e7a3;
+  text-decoration-thickness: 3px;
 }
