+++ website/src/pages/initiatives.js
@@ -4,7 +4,7 @@ import styles from './initiatives.module.css';
 import CommunityCard from '../components/CommunityCard';
 import { FixedSizeList as List } from 'react-window';
 import MeetingCard from '../components/PastMeetingCard';
-import { pastMeetings, upcomingMeetings } from '../../communityData/meetings';
+import { pastMeetings, upcomingMeetings } from '../../websiteData/meetings';
 
 const Community = () => {
   return (
